+++ _get.sh
@@ -1,0 +1,10 @@
+flutter pub get
+cd demo
+flutter pub get
+cd ..
+cd example
+flutter pub get
+cd ..
+cd packages/pluto_grid_export
+flutter pub get
+cd ../..