+++ lib/src/ui/columns/pluto_column_title.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:pluto_grid/pluto_grid.dart';
+import 'package:pluto_grid_plus/pluto_grid_plus.dart';
 
 import '../ui.dart';
 
@@ -224,8 +224,8 @@ class PlutoGridColumnIcon extends StatelessWidget {
     this.descendingIcon,
     this.successColor = Colors.green,
     this.errorColor = Colors.red,
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -267,8 +267,7 @@ class _DraggableWidget extends StatelessWidget {
     required this.stateManager,
     required this.column,
     required this.child,
-    Key? key,
-  }) : super(key: key);
+  });
 
   void _handleOnPointerMove(PointerMoveEvent event) {
     stateManager.eventManager!.addEvent(PlutoGridScrollUpdateEvent(
@@ -326,11 +325,10 @@ class _SortableWidget extends StatelessWidget {
   final Widget child;
 
   const _SortableWidget({
-    Key? key,
     required this.stateManager,
     required this.column,
     required this.child,
-  }) : super(key: key);
+  });
 
   void _onTap() {
     stateManager.toggleSortColumn(column);
@@ -362,8 +360,7 @@ class _ColumnWidget extends StatelessWidget {
     required this.stateManager,
     required this.column,
     required this.height,
-    Key? key,
-  }) : super(key: key);
+  });
 
   EdgeInsets get padding =>
       column.titlePadding ??
@@ -438,8 +435,7 @@ class _ColumnWidget extends StatelessWidget {
 class CheckboxAllSelectionWidget extends PlutoStatefulWidget {
   final PlutoGridStateManager stateManager;
 
-  const CheckboxAllSelectionWidget({required this.stateManager, Key? key})
-      : super(key: key);
+  const CheckboxAllSelectionWidget({required this.stateManager, super.key});
 
   @override
   CheckboxAllSelectionWidgetState createState() =>
@@ -515,8 +511,7 @@ class _ColumnTextWidget extends PlutoStatefulWidget {
     required this.stateManager,
     required this.column,
     required this.height,
-    Key? key,
-  }) : super(key: key);
+  });
 
   @override
   _ColumnTextWidgetState createState() => _ColumnTextWidgetState();